finished MVP!
everything actually works now on a test chat! there are still few bugs to be concerned about:
 - the bot pings itself. not a big deal, but still, looks unpolished
 - currently, every chat gets a different csv. this could be problematic though, since
  a) the name of the chat can be emoji, which could case problems in a filepath
  b) the values inside the csv are going to be outdated with any change
 however, it should be considered that this code might be called from multiple chats at the same time, so race conditions should be checked (a random base64 can be used as a temporary filename and then deleted, for example
 - the bot doesn't understand commands and it only works via a single .env parameter (the name of the chat). I can vision it as pinging the ping bot triggers the "ping all" functionality (with some confirmation and in-chat logs (NO TELEGRAM ID's IN THE SYS LOGS ARE ALLOWED!!)
