+++ src/obtaining_csv_with_users.rs
@@ -1,6 +1,6 @@
-use crate::find_chat_by_name;
-use grammers_client::{Client, types::PackedChat};
-use log::{error, info};
+use grammers_client::Client;
+use grammers_client::types::Chat;
+use log::info;
 use serde::{Deserialize, Serialize};
 use std::error::Error;
 use tokio::time::{Duration, sleep};
@@ -20,26 +20,13 @@ impl MyUser {
     }
 }
 
-pub async fn generate_csv(client: &Client, chat_name: &str) -> Result<(), Box<dyn Error>> {
-    // let chat_name = env::var("CHAT_NAME_TO_PARSE").unwrap();
-    // 3. Resolve chat username into PackedChat
-    let packed_chat = match find_chat_by_name(client.iter_dialogs(), chat_name).await {
-        None => {
-            error!("Can't find chat with name {chat_name}");
-            panic!();
-        }
-        Some(ch) => {
-            info!("Found chat {}", ch.name());
-            ch.pack()
-        }
-    };
-
+pub async fn generate_csv(client: &Client, chat: &Chat) -> Result<(), Box<dyn Error>> {
     // 4. Export participants to CSV
-    let csv_filepath = format!("{chat_name}.csv");
+    let csv_filepath = format!("{chat_name}.csv", chat_name = chat.name()); // TODO: make it id, not name, to avoid emojis being a problem
     let mut wtr = csv::Writer::from_path(&csv_filepath)?;
 
     // the csv header is used automatically when serializing from a struct
-    let mut iter = client.iter_participants(packed_chat);
+    let mut iter = client.iter_participants(chat);
     while let Some(part) = iter.next().await? {
         let user = &part.user;
         let my_user = MyUser {
@@ -56,7 +43,7 @@ pub async fn generate_csv(client: &Client, chat_name: &str) -> Result<(), Box<dy
     return Ok(());
 }
 
-pub fn get_pings_from_file(filepath: &str) -> Result<String, Box<dyn Error>> {
+pub fn parse_pings_from_file(filepath: &str) -> Result<String, Box<dyn Error>> {
     // 5. Read CSV and ping each user
     let mut result_message = String::new(); // the final message with space-separated mentions
     let mut rdr = csv::Reader::from_path(filepath)?;
